Bug fixing
This bug crashes typescripts compilation: error TS2540: Cannot assign to 'origin' because it is a constant or a read-only property.

As origin is a read only field and cannot  be changed, and this function at all returns only combined value, I commented the part of assignment of a value to a read only field.