Features/35941 refapp deploy optimizations (#233)
* Deleted build/gulp-tasks/config.js

a) Typescript configurations settings moved to separate tsconfig.json files into Build/tsconfigs folder
b) filenames, globs, paths, watchesTemplate params have no usage references
c) karma config data moved to the karma config in Build/Karma.conf.js file, but also:
- excluded default Karma server port;
- excluded coverage plugin as it was not working during deploy and need to be investigated;
- excluded HTML reporter as it has dependencies with very critical bugs and we have JUnit to see file reports;

* Deleted build/gulp-tasks/gulp-auto-reload-config.js

This file used to reload GULP configs, since we don't use GULP we don't need this file anymore

* Deleted build/gulp-tasks/helpers.js

These helpers included supporting operations for GULP, like logging, file operations, also typescript compilation and running unit tests. Unit tests were moved to Cake, typescripts compilation as well. For eachBladeStreams and globFromRootFolder functions have not found any usage references.

* Deleted build/gulp-tasks/karma.conf.js

It was predefined karma config, which took data from other config files. Since we stick all karma configs to Build/karma.conf.js file, we don't need this one

* Deleted build/gulp-tasks/readme.md

Description for GULP. Since we don't use gulp anymore, this description is redundant.

* Deleted  build/gulp-tasks/scripts.js

This file was used to generate javascript files from typescripts via GULP. Since we generate javascript in Cake, and have a task for typescript generation, this file is not needed anymore

* Deleted build/gulp-tasks/server.js

Haven't found usage references

* Deleted  build/gulp-tasks/styles.js

Have not found usage references, and styles compiling via CSS packages

* Deleted build/gulp-tasks/unitTests.js

Running typescripts unit tests with Karma was designed in Cake instead of GULP

* Deleted build/gulp/config.js

Another part of Karma configuration, was moved to the major build/Karma.con.js file. "documentationSettings" param seems to be unused, haven't found any references

* Deleted build/gulp/documentation.js

I have not found any references for this file usage. It seems to be called manually (if was). But even manually, it does not work. If we need documentation implementation, we can add this in Cake.

* Deleted build/gulpfile.js

Major gulp file, since we don't use Gulp we don't need this file

* Deleted build/readme.md

Description for GULP, since we don't use gulp, this file is redundant.

* Updated build/package.json

- Deleted unused packages;
- Used packages were updated as much as possible
- karma-chrome-launcher kept for debugging purposes;
- karma-html-reporter was removed because of critical issues, but kept junit reporter

* Created build/tsconfigs/orckestra.json

Configuration file with compilation settings of Typescript for release

* Created build/tsconfigs/orckestra_tests.json

Configuration file with compilation settings of Typescript for tests (to be tested). It excludes  App.ts

* Created build/tsconfigs/unittests.json

Configuration file with compilation settings of Typescript unit tests to be run with Karma

* Updated build/karma.conf.js

Karma config file with all values in the same place. Were excluded default values, htmlReporter (as it has critical dependencies and we can use junit reporter to see a log file)

* Updated .gitignore

To ignore the javascript file generated for unit tests (to be tested). This file will be deleted after the build but added this rule just for case.

* Fixed CheckoutService.spec.ts

After Jasmine update, this test starts from time to time fails. It was because of the different order of tests running in the last Jasmine version. To fix the errors, it was added initial value for some tests, so they are not dependent on running order anymore.

* Fixed MockJqueryEventObject.ts

Gulp-Typescript plugin very often skips errors, which should appear, in particular inheritance and not-existing-reference errors. After changing the compiler to Microsoft Typescript Compiler inheritance error starts to appear, here is a little bug fix

* Updated webrequests.cake

To this helper was added the GetServerFileInfo method which returns server file modified timestamp and size in bytes, without reading the whole file, just headers.

* Created typescripts.cake helper

To Cake was added a CompileTypeScripts helper. Since we compile typescripts for different purposes, this helper was separated. It uses Microsoft Typescripts compiler

* Updated webrequests.cake helper

Method optimization to return more detailed information

* Updated install.cake

Updating logic of C1 File downloading. In case we have local a file with the same size and changed timestamp, then not to reload it as it takes from 10 s to a few minutes.
Changed a path of a cached C1 file, from .temp to .cache since a .temp folder has to be cleared.

* Updated build.cake

- added Microsoft Typescript Compiler package for typescript compilation. Microsoft Typescript compiler is a more reliable tool since Gulp-Typescript has some known issues. Also since Gulp will not be used during a deploy process, it will be not possible to use the Gulp-Typescript plugin.
- added Cake Karma addon;
- clearing .temp folder before building;
- deleted tasks referred to GULP.

- added task Compile-Typescripts to compile typescripts for release;
- added task Karma-Tests to run Karma unit tests
- added task Karma-Debug to run debugging of Karma unit tests
- refactoring

* Updated Orckestra.Composer.Website.csproj

Excluded not existed project file

* Fixed Website - CD.pubxml

Changing a path to a relative one

* Fixed AutocompleteSearchBoxController.ts

Replacing references from App.ts to certain ones

* Fixed CartRepository_AddLineItemAsync.cs

Quantity cannot be zero, generated 0 value caused build crash. Replaced to a static value as this UT is not designed to test quantity and Microsoft recommends do not create dependencies in UT

* Updated project README.md
Updated general RefApp description

* Updated Orckestra.Composer.Website.csproj

App.configs excluded from republishing from Visual Studio