Bugs fixes and some refactoring (#2)
#AB37061

- Added some null checks
- Bug fixes
- Refactoring

* Update CartViewModelFactory.cs

Added a null check

* Update RecurringOrderCartsViewService.cs

Added a comparison using ignore-case

* Update SearchUrlProvider.cs

Bugfix. SortedDicrionary replaced to a List because the name value is not unique and that can lead to exceptions

* Update SitemapProvider.cs

Added a null check

* Update ProductPriceService_GetProductPriceAsync.cs

Removed unused code, added param check

* Update StoreLocatorViewService_GetStoreLocatorViewModelAsync.cs

Wrong method name since we validate property of param, but not a param.  It is confusing since we except ArgumentException

* Update StoreViewService_GetStoreViewModelAsync.cs

Wrong method name since we validate property of param but not a param

* Update ProductPriceService_GetProductPriceAsync.cs
Replaced index to a LINQ call